More abstraction; Server binaries for proxy & neo4j
Break the neo4j and proxy servers into their own binaries,
this means that endpoint.py is now abstracted so that the
flask app may be configured and managed in a cleaner fashion.

It's still not perfect... we're modifying the backend
variable of the flask app directly as a global, but
it's something!

Signed-off-by: Eric Windisch <eric@windisch.us>